Added the odds(array) and evens(array) definitions
Propose two more functions to be added to the library. One returns all odd numbers from list, the other returns all even numbers. I know that this is rather trivial, but I always wanted this to be a library function, rather then having to type my own helper methods for it over and over again...